mountains, clouds, and trees
4. Render background items [2 mark]

From part 4_scrollingBackground, copy and paste the arrays `trees_x`, `clouds`, and `mountains`.

Copy and paste the `for loops` for each of these items into the `draw` function.

Now we are going to refactor this code.

Create three new functions: `drawClouds`, `drawMountains`, `drawTrees`.

Cut and paste each `for loop` into the relevant function.

In the `draw` function, call each of the three new functions (e.g. `drawClouds`).

You should now see all of the above items drawn to the screen.