canyons and collectables
5. Render canyons and collectables [2 marks]

From 4_scrollingBackground, copy and paste the arrays `canyons`, `collectables`.

Copy and paste the `for loops` for each of these items into the `draw` function.

We are also going to refactor this code but we will do things a little differently.

Cut and paste only the body of each `for loop` into the functions `drawCanyon` and `drawCollectable`.

Now replace the references to the arrays with the argument provided by the function (e.g. `canyons[i].posX` becomes `t_canyon.posX`).

Finally we need to call these functions passing each canyon or collectable as a parameter.

Do this inside the body of each `for loop` (e.g. `drawCanyon(canyons[i]);`).

You should now see canyons and collectables rendered. HINT: check the draw order is correct.