implement collectables interaction
7. Implement collectables interaction [2 marks]

From part 3b_canyonsAndCoins, copy the conditional statement which checks whether a player is within range of a collectable and paste it into the function `checkCollectable`.

As you did with the `drawCollectable` function, replace the references to `collectable` with references to `t_collectable`.

Write the code to call `checkCollectables` just below the call to `drawCollectables`. It should use exactly the same format.

We now need to make sure that the collectable is not drawn once it has been found. Add a conditional statement to the same `for loop` which prevents both functions being called if that collectable's `isFound` property is `true`.

Test your code to check that collectables disappear when the game character collides with them.

You'll notice that once the screen starts scrolling the collectable items stop functioning as expected. This is because `gameChar_x` only represents where the character appears on the screen, not where they are in the game world.

The variable `gameChar_world_x` shows where the game character is in the game world. Replace `gameChar_x` in the `checkCollectable` function with `gameChar_world_x` to make the collectable items work with scrolling.