canyons interaction
8. Implement canyons interaction [1 mark]

From part 3b_canyonsAndCoins, copy and paste the conditional statement which makes the character plummet into the `draw` function.

Copy and paste the conditional statement which detects whether the player is over a canyon into the function `checkCanyon`.

Adapt the code to use the argument `t_canyon`.

Adapt the same code to use `gameChar_world_x` instead of `gameChar_x`.

Call the function from the same `for loop` which draws the canyons, passing each canyon as an argument in the same way as you did for `drawCanyon`.