add a flagpole
2. Add a flagpole [1 mark]

We need to add an end to your level. I have chosen a flagpole but you can choose according to the theme of your game.

Initialise an object called `flagpole`; it should at least have the properties `x_pos` and `isReached`.

Set `isReached` to `false` and `x_pos` to a world position at the very end of your level.

Create a function called `renderFlagpole` and call this from the draw function.

Complete the function to draw your flagpole in two states. One for when `isReached` is false, and one for when it is `true`.